update temperature to a uint64, update storage layout for gas efficency

+++ contracts/beanstalk/init/InitializeDiamond.sol
@@ -189,7 +189,7 @@ contract InitializeDiamond {
         s.sys.weather.thisSowTime = type(uint32).max;
         s.sys.weather.lastSowTime = type(uint32).max;
         s.sys.weather.morningDuration = 600; // 10 minutes
-        s.sys.weather.morningControl = uint256(1e18) / 240; // 1 / 240 = 0.004166666667
+        s.sys.weather.morningControl = uint128(1e18) / 240; // 1 / 240 = 0.004166666667
         s.sys.extEvaluationParameters.minSoilIssuance = MIN_SOIL_ISSUANCE;
     }
